+++ dn/kit-viewer.hpp
@@ -117,7 +117,7 @@ struct Viewer : Widget {
 	explicit Viewer(Kit &kit);
 	~Viewer() override;
 
-	void measure(Kit &kit, float max_w, float max_h) override;
+	void measure(Kit &kit, int max_w, int max_h) override;
 	void arrange(Kit &kit, Rect alloc) override;
 	void paint(Kit &kit) const override;
 	void prepare(Kit &kit) override;
